feat: implement chat functionality with tool integration and UI components
